Remove match and use competition in its place (Closes #1)
Match is a much more fitting name for an active game. Competition is too vague and thus causes some confusion when navigating the codebase. "Arena" represents the actual game class itself, and contains the majority of game logic and rules for the game. Actual game arenas will be referenced to as maps.